+++ packages/cloud-cognitive/src/components/CreateTearsheet/CreateTearsheet.test.js
@@ -15,7 +15,7 @@ import { CreateTearsheet } from './CreateTearsheet';
 import { CreateTearsheetStep } from './CreateTearsheetStep';
 import { CreateTearsheetSection } from './CreateTearsheetSection';
 import uuidv4 from '../../global/js/utils/uuidv4';
-const tearsheetBlockClass = `${pkg.prefix}--tearsheet-create`;
+const createTearsheetBlockClass = `${pkg.prefix}--tearsheet-create`;
 
 const rejectionErrorMessage = uuidv4();
 const onCloseFn = jest.fn();
@@ -67,18 +67,20 @@ const renderCreateTearsheet = (
       <p>Child element that persists across all steps</p>
       <CreateTearsheetStep
         onNext={rejectOnNext ? onNextStepRejectionFn : onNext}
-        title={step1Title}>
+        title={step1Title}
+        formLegendText={step1Title}>
         step 1 content
         <button type="button" disabled>
           Test
         </button>
         <input type="text" />
       </CreateTearsheetStep>
-      <CreateTearsheetStep title={step2Title}>
+      <CreateTearsheetStep title={step2Title} formLegendText={step2Title}>
         step 2 content
       </CreateTearsheetStep>
       <CreateTearsheetStep
         title={step3Title}
+        formLegendText={step3Title}
         onNext={rejectOnSubmitNext ? finalStepOnNextRejectFn : finalOnNextFn}>
         step 3 content
       </CreateTearsheetStep>
@@ -103,8 +105,7 @@ const renderEmptyCreateTearsheet = ({ ...rest }) =>
 const renderSingleStepCreateTearsheet = ({ ...rest }) =>
   render(
     <CreateTearsheet onRequestSubmit={onRequestSubmitFn} {...rest}>
-      <p>Child element that persists across all steps</p>
-      <CreateTearsheetStep title={step1Title}>
+      <CreateTearsheetStep title={step1Title} formLegendText={step1Title}>
         step 1 content
       </CreateTearsheetStep>
     </CreateTearsheet>
@@ -138,11 +139,13 @@ describe(CreateTearsheet.displayName, () => {
   it('renders the CreateTearsheet component', () => {
     const { container } = renderCreateTearsheet({
       ...defaultProps,
-      'data-test-id': dataTestId,
+      'data-testid': dataTestId,
     });
     screen.getByTestId(dataTestId);
     screen.getAllByText(title);
-    expect(container.querySelector(`.${tearsheetBlockClass}`)).toBeTruthy();
+    expect(
+      container.querySelector(`.${createTearsheetBlockClass}`)
+    ).toBeTruthy();
     expect(ref.current).not.toBeNull();
   });
 
@@ -153,11 +156,11 @@ describe(CreateTearsheet.displayName, () => {
     const cancelButtonElement = screen.getByText(cancelButtonText);
     click(nextButtonElement);
     const createTearsheetSteps = container.querySelector(
-      `.${tearsheetBlockClass}__content`
+      `.${createTearsheetBlockClass}__content .${carbon.prefix}--form`
     ).children;
     expect(
       createTearsheetSteps[1].classList.contains(
-        `.${tearsheetBlockClass}__step--visible-section`
+        `.${createTearsheetBlockClass}__step--visible-section`
       )
     );
 
@@ -195,11 +198,11 @@ describe(CreateTearsheet.displayName, () => {
     });
     click(nextButtonElement);
     const tearsheetChildren = container.querySelector(
-      `.${tearsheetBlockClass}__content`
+      `.${createTearsheetBlockClass}__content  .${carbon.prefix}--form`
     ).children;
     expect(
       tearsheetChildren[2].classList.contains(
-        `.${tearsheetBlockClass}__step--visible-section`
+        `.${createTearsheetBlockClass}__step--visible-section`
       )
     );
     rerender(
@@ -207,11 +210,10 @@ describe(CreateTearsheet.displayName, () => {
         {...defaultProps}
         open={false}
         onRequestSubmit={onRequestSubmitFn}>
-        <p>Child element that persists across all steps</p>
-        <CreateTearsheetStep title={step1Title}>
+        <CreateTearsheetStep title={step1Title} formLegendText={step1Title}>
           step 1 content
         </CreateTearsheetStep>
-        <CreateTearsheetStep title={step2Title}>
+        <CreateTearsheetStep title={step2Title} formLegendText={step2Title}>
           step 2 content
         </CreateTearsheetStep>
       </CreateTearsheet>
@@ -326,7 +328,7 @@ describe(CreateTearsheet.displayName, () => {
   it('should not render any CreateTearsheet steps when there are no TearsheetStep components included', () => {
     const { container } = renderEmptyCreateTearsheet(defaultProps);
     const createTearsheetSteps = container.querySelectorAll(
-      `.${tearsheetBlockClass}__step`
+      `.${createTearsheetBlockClass}__step`
     );
     expect(Array(...createTearsheetSteps)).toStrictEqual([]);
   });
@@ -346,11 +348,11 @@ describe(CreateTearsheet.displayName, () => {
     const backButtonElement = screen.getByText(backButtonText);
     click(backButtonElement);
     const tearsheetChildren = container.querySelector(
-      `.${tearsheetBlockClass}__content`
+      `.${createTearsheetBlockClass}__content`
     ).children;
     expect(
       tearsheetChildren[0].classList.contains(
-        `.${tearsheetBlockClass}__step--visible-section`
+        `.${createTearsheetBlockClass}__step--visible-section`
       )
     );
   });
@@ -392,7 +394,7 @@ describe(CreateTearsheet.displayName, () => {
     const { click } = userEvent;
     click(screen.getByText(viewAllToggleLabelText));
     const viewAllToggleElement = container.querySelector(
-      `#${tearsheetBlockClass}__view-all-toggle`
+      `#${createTearsheetBlockClass}__view-all-toggle`
     );
     setTimeout(() => {
       expect(viewAllToggleElement).toBeChecked();