+++ packages/cloud-cognitive/src/global/js/use/index.js
@@ -1,2 +1,0 @@
-export { useWindowScroll } from './useWindowScroll';
-export { useWindowResize } from './useWindowResize';